feat: Show chat list from local storage
This commit fetches the list of chats from the database through
ChatRepository class. The repository reads the chats from the local
database and then converts entity response to value object,
independent of any framework/representation. The commit also setups a
handler for clicking of chat item, which prints the chat to logs.