feat: Show full name in toolbar when ChatDetailFragment is shown
This commit uses a separate toolbar for flexibility. It also setups the
toolbar with the navigation controller in MainActivity.
MainActivity listens to destination changes in the navigation graph and
sets the toolbar title to chat user's full name in ChatDetailFragment
through the passed bundle argument to the fragment.
It also uses fragment tag instead of FragmentContainerView due to a bug
when it's used with NavHostFragment